// src/api/generated/models/productListItemResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { ProductDefaultVariantSummaryDto } from './productDefaultVariantSummaryDto';
import type { ProductListItemResponseDtoStatus } from './productListItemResponseDtoStatus';
import type { ProductRelationSummaryDto } from './productRelationSummaryDto';

export interface ProductListItemResponseDto {
  /**
     * @minLength 26
     * @maxLength 26
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  id: string;
  name: string;
  slug: string;
  status: ProductListItemResponseDtoStatus;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  supplier?: ProductRelationSummaryDto | null;
  /** @nullable */
  publisher?: ProductRelationSummaryDto | null;
  categories: ProductRelationSummaryDto[];
  authors: ProductRelationSummaryDto[];
  /** @nullable */
  defaultVariant?: ProductDefaultVariantSummaryDto | null;
  /** @nullable */
  minPrice?: string | null;
  /** @nullable */
  maxPrice?: string | null;
  variantCount: number;
  activeVariantCount: number;
  optionCount: number;
  createdAt: string;
  updatedAt: string;
}